requirements/research.txt
arxiv
docling<=2.38.1
docling-core<=2.38.2
exa-py
google-search-results
gradio>=5.0.0
json5
markdown
mcp
modelscope
neo4j
openai
pandas
Pillow
python-dotenv
requests
rich
sentence-transformers
langchain
